Correct addBookToLibrary method in LibraryDAO

+++ src/main/java/codecooler/michal/com/LibraryDAO.java
@@ -1,7 +1,7 @@
 package codecooler.michal.com;
 
 public interface LibraryDAO {
-    void addBookToLibrary(int ISBN, int author_id, String title, String publisher_id, int publication_year, int price);
+    void addBookToLibrary(long ISBN, int author_id, String title, String publisher_id, int publication_year, int price);
     void editBookData(String title, int publication_year, int price, int author_id);
     void removeBook(long author_id);
     void findBookByAuthorSurname(String surname);
